third_party/blink/web_tests/external/wpt/FileAPI/reading-data-section/filereader_events.any-expected.txt

This is a testharness.js-based test.
[FAIL] events are dispatched in the correct order for an empty blob
  assert_equals: Expected load event, but got progress event instead expected "load" but got "progress"
Harness: the test ran to completion.
